A quick fix for rake schema
Now there is a ternary in Generator::Base when the name attribute is being initialized. The error was occurring because
in the schema.rake task the argument for name is nil...Generator::Schema.new(nil).rebuild. The private method in Base, normalize_name
which underscores the name was throwing NoMethodError. Now the method normalize_name is not used if the name is nil.