No more deprecated `-i` pg_dump option
This option is deprecated since [PostgreSQL 8.4][1]. So it does not make sense to keep it.

`pg_dump` bundled with PostgreSQL 9.5 even fails to run with this option. That makes impossible to run `rake db:schema:dump`. It looks like this:

```
> bundle exec rake db:schema:dump
pg_dump: invalid option -- i
Try "pg_dump --help" for more information.
```

[1]: http://www.postgresql.org/docs/8.4/static/app-pgdump.html