Add authorization helpers
Similar to `Pliny::Helpers::Serialize` authorization helpers define public interface that might be used to set and call an authorizer on specific or all endpoints.

Examples

```ruby
module Authorizers
  class Token
    def initialize
      @token = 'foo'
    end

    def authorized?(request)
      @token == request.params['access_token']
    end
  end
end

module Endpoints
  class Foo < Base
    authorizer Authorizers::Token

    before { authorize! }

    get do
    end
  end
end
```

Authorizers are PORO that must respond to `#authorized?` predicate method. The method receives a request object as a parameter and returns true or false depending on results of the authorization.