By default, log canonical lines with just default context
In retrospect, it seems to make sense to log canonical lines with just
default logging context so that it'll include an app name and any other
information that a user has deemed critical. There is some trade off
here because the information included by the canonical logger becomes
less explicit, but that'll probably be okay.