Sort by filename when generating INSERTs
Previously, the only thing keeping the output of .sql generation was
the insert order of tuples happening to the recall order...something
that can easily de-synchronize with multiple developers that may be
writing migrations at the same time.

Instead of getting the tuples in any order, sort them by their
content.