Allow activesupport 5 or 6
We have a very minimal dependency on active support - limited to the
inflection extensions in a generator, and some time extensions in a
test. So it seems safe to continue allowing 5, while opening up 6. We'll
lock out 7 because it doesn't exist, and we don't know what, if any,
breaking changes it will bring.