Upgrade Rake and rack-test
Rake is way out of date, and we really on use it as a binary, so it's
pretty low risk.

rack-test is purely a dev resource for pliny, even though it's in the
template Gemfile for `pliny new`. But it's got no version restraints
there, so again, very low risk.