Ensure newline characters are removed from log strings
Pliny logging follows the "logfmt" pattern[1]. logfmt is unclear on what should happen with newline characters, but given it's line-based it doesn't seem to make sense to include them in the body of the log line. I couldn't decide between removing, replacing with a space character, or escaping. In the end I went with escaping since it makes it clear to a viewer that a newline _was_ there.

[1] https://brandur.org/logfmt