Do serializer configuration properly
Looking at the sinatra extensions guide[1] it looks like you're meant to use `set` to set this kind of class-level state.

The value has to be initialised to nil otherwise you get `NoMethodError: undefined method `serializer_class'`, I guess because the accessor method is only created when `set` is called.

[1] http://sinatrarb.com/extensions.html#setting-options-and-other-extension-setup