Add serializer DSL correctly
Following the sinatra extensions guide[1] it looks like this is how you're meant to do it.

Unfortunately this results in a breaking change in pliny because pliny apps will have to switch their endpoint classes to use `register` instead of `helper` (see the change in the template in this commit). But it fixes a compatibility issue between pliny and sinatra 2.1 where the latter has switched its internal behaviour, using `prepend` rather than `include` for helpers, meaning the `self.included` hook is no longer being called.

[1] http://sinatrarb.com/extensions.html#extending-the-dsl-class-context-with-sinatraregister