help RuboCop pick the correct config
The problem is that RuboCop by default picked the inner most config
file. This means that if we want to use RuboCop for Pliny itself with a
.rubocop.yml config in the project root, we always need to append `-c
.rubocop.yml`. If you don't do this you are greeted with a lot of
errors. Which is especially annoying if you have RuboCop integrated in
your editor. With this change it works out of the box. With a minimal
trade-off by renaming the template.